Update Inv Models and Controller

+++ Models/InvoiceItem.cs
@@ -9,9 +9,9 @@ namespace invoice_system_backend.Models
 
         public string Description { get; set; }
 
-        public float Quantity { get; set; }
+        public decimal Quantity { get; set; }
 
-        public float Price { get; set; }
+        public decimal Price { get; set; }
 
         public int InvoiceId { get; set; }
     }
